+++ jest.config.ts
@@ -6,10 +6,10 @@ export default {
   coverageThreshold: {
     // 所有文件总的覆盖率要求
     global: {
-      branches: 80,
-      functions: 80,
-      lines: 80,
-      statements: 80,
+      branches: 70,
+      functions: 70,
+      lines: 70,
+      statements: 70,
     },
   },
   // 用排除法，解决lint-stage增量时的问题
@@ -38,4 +38,5 @@ export default {
       },
     ],
   },
+  maxWorkers: 1,
 };