+++ test/decorator.test.ts
@@ -14,7 +14,11 @@ describe('decorator', () => {
       host: 'xxx.xxx.xxx',
       user: 'xxxx',
       password: 'xxxxx',
-      database: 'xxxx'
+      database: 'xxxx',
+      autoSync: {
+        enable: false,
+      }
+
     })
   });